+++ src/tk/CfgUtil.hh
@@ -1,6 +1,6 @@
 //
 // CfgUtil.hh for pekwm
-// Copyright (C) 2021-2023 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2021-2025 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
@@ -18,6 +18,8 @@ namespace CfgUtil {
 	void getThemeDir(const CfgParser::Entry* root,
 			 std::string &theme_dir, std::string &theme_variant,
 			 std::string &theme_path);
+	bool lookupThemeVariant(const std::string &theme_file,
+				std::string &variant);
 	void getIconDir(const CfgParser::Entry* root, std::string &icon_dir);
 	void getScriptsDir(const CfgParser::Entry* root,
 			   std::string &scripts_dir);